Render screenshots by capturing the viewport instead of the page content
Uses the new 'mode' option from the PhantomJS render API so that
screenshots get captured based on the viewport size definition and not the
content size. Indeed, some HTML presentation frameworks use the position
and transform CSS properties that leads to having the visible content
larger than the viewport.

Increases the delay for the resize event to complete.