Upgrade shrinkwrap
This upgrades npm-shrinkwrap.json to lockfile version 2, which is the format introduced by npm v7. This fixes chrome not being downloaded when installing decktape as a dependency in a project using npm v7+. See https://github.com/puppeteer/puppeteer/issues/6586#issuecomment-958379915. The new format is backward compatible so users of npm v6 and lower should not be impacted.